marker shows in_flock, speed up or slow down

+++ abm/flock_sim.py
@@ -2,36 +2,48 @@ from time import sleep
 # from datetime import datetime
 import bike_agent as model
 from router import refine, Router
+import argparse
+
+
+parser = argparse.ArgumentParser(description='grab adscriptions from medline')
+parser.add_argument('--init', type=int, default=50,
+                    help="set to 0 to resume run")
+args = parser.parse_args()
+
 
 model.connect('mydb')
 
-# drop all bikes
-for b in model.Bike.objects.all():
-    b.delete()
+if args.init > 0:
+    # drop all bikes
+    for b in model.Bike.objects.all():
+        b.delete()
 
-h = 0.0
-# init source, target, speed for this many bikes
-for n in range(100):
-    b = model.Bike()
-    b.speed = 30
-    b.random_ride(ne_lat=19.461332069967366,
-                  ne_lng=-99.09204483032227,
-                  sw_lat=19.40467336236742,
-                  sw_lng=-99.17787551879884,
-                  min_len=5, max_len=10)
+    h = 0.0
+    # init source, target, speed for this many bikes
+    for n in range(args.init):
+        b = model.Bike()
+        b.speed = 20
+        b.random_ride(ne_lat=19.461332069967366,
+                      ne_lng=-99.09204483032227,
+                      sw_lat=19.40467336236742,
+                      sw_lng=-99.17787551879884,
+                      min_len=5, max_len=10)
 
-    b.route = refine(b.route, b.speed)
-    b.save()
-    sleep(0.1)
-    print "creada %s" % b
+        b.route = refine(b.route, b.speed)
+        b.save()
+        sleep(0.1)
 
 
 def status_count():
     solo = float(model.Bike.objects(status="solo").count())
-    flocking = float(model.Bike.objects(status="flocking").count())
+    in_flock = 0
+    for b in model.Bike.objects.all():
+        if b.in_flock(100):
+            in_flock += 1
+    # flocking = float(model.Bike.objects(status="flocking").count())
     all = float(model.Bike.objects.count())
     return {'solo': "%0.2f" % (solo / all),
-            'flocking': flocking / all}
+            'in': "%0.2f" % (in_flock / all)}
 
 
 def mean_distance():
@@ -63,13 +75,17 @@ while model.Bike.objects.count() > 0:
 
                     b.route = refine(router.route, b.speed)
                     b.status = 'flocking'
+                    b.speed = 30
                     b.save()
             else:
-                if b.status == 'flocking':
+                if not b.in_flock(100):
                     router = Router(points=[b.point['coordinates'],
                                             b.destination['coordinates']])
                     b.route = refine(router.route, b.speed)
                     b.status = "solo"
-                    b.save()
+                    b.speed = 20
+                else:
+                    b.speed = flock.mean_speed
+                b.save()
 
     print status_count(), mean_distance()
